requirements.txt
mistralai
streamlit
faiss-cpu
pypdf
mistralai